Finally Got Jekyll to Build!
After several weeks of failures to build, jekyll is finally working again.  Sort of.

Could be attribted to:
- new install of xcode-select cmd line tools
- update of all gems, including an update of the listen gem (finally?) that works with Ruby 3.0.0

Jekyll does seem to be building properly, but I can't get it to serve locally.  Will check this out later.

Also added "itworks_2021-02-07.pdf" to the working folder to show the iTerm2 buffer as of this successful build.  Hopefully can compare to the command line entries I submitted as part of bugs, to see what's different.