bundle add webrick
Thanks to https://github.com/jekyll/jekyll/issues/8531 I saw that Jekyll wasn't serving because I was missing the webrick gem.

Now that that's added, everything seems to be working!

In the future I will work with gems inside of the repo and I am now with gems in the vendor directory.  Just seems better to keep everything contained.