Placing assets for watson-assistant-chatbot in main directory
I've had this problem since creating a portfolio collection of duplicate assets.  (To be able to see assets in the _portfolio folder as I edit, but to also publish them to the correct location.)

By placing assets in the root /assets folder, the locations don't change when the site is built to _site.

I means I can't have a single folder with the built portfolio page as index.html and assets in an 'assets' folder, next to it, but that's okay.